+++ tscurvebuffer.h
@@ -68,7 +68,7 @@ signals:
     void updateAverageData(int avgTempIn, int avgTempOut, int avgDo, int ChD);
 public slots:
     void updateAvData(int avgTempIn, int avgTempOut, int avgDo, int ChD);
-    void appendData(QVector<double> volume, QVector<double> tempin, QVector<double> tempout);
+    void appendData(QVector<int> volume, QVector<int> tempin, QVector<int> tempout);
 private:
 
     QVector<int> ts_volume;
